move marks into its own test
